Moved interpolation function into netcdf_data
